requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-filter==23.4
Pillow==10.1.0
markdown==3.5.1
